renamed ReactTable to BaseTable
